WIP: Add lifting-cast-config web app served by Flask
Does this work?
The goal is for the Flask app in lc.py to be started by main.py in a
daemon thread. The Flask app should
* Serve the web app (in lifting-cast-config-app/public) on
  http://localhost:5000 (default development server port), to be
  available in a web browser on the pi or another computer on the
  network.
* Receive the LiftingCast config data (meet id, password, platform id)
  from the web app. It should then set the corresponding variables and
  set `configured` to `True`.

DRL should simply see the new values of those variables as it checks
them as the event loop goes around.

I have tested the Flask app hosting the web app but does the DRL
integration actually work?